api/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
dnspython==2.6.1
environs==11.0.0
fastapi==0.112.0
fastapi_cors==0.0.6
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.31.2
MarkupSafe==2.1.5
marshmallow==3.21.3
msgpack==1.0.8
numpy==2.0.1
packaging==24.1
pandas==2.2.2
psutil==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
pyzmq==26.1.0
requests==2.32.3
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
strawberry-graphql==0.237.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
Werkzeug==3.0.3
zope.event==5.0
zope.interface==7.0.1
